use Left zone maps for both hemispheres
This is being done to test if the left LPFC removal causing a larger change in variance is due to the mask definition by the zones. 

We will use the left watershed zones for both the left and the right in this following analysis. This may change back to the original -- to change back to original analysis comment in line 106 and out Line 107